Add MultiSelectListPreference for Auto freeze tags to Settings.
Create list dynamically based on existing tabs. This is done while
creating the preferences: I'm not sure this is the nicest way to
do it, but couldn't think of anything better.

Store the list in a new AUTO_FREEZE_TAGS Shared Preference (in HailData).

Add and use 'ic_outlook_local_offer' as image for setting (because
it looks like a tag - and probably is).